Add files via upload
Updated on 7/11/2023  - Here I have tried incorporating the Page Object Module (POM) to perform the Front-end testing. The code is just the same as the previous ones, but I have put all the three testcases in one place and used classes for each webpage and several methods inside it to perform the required tasks efficiently.